+++ src/JTest.Core/Output/ITestCaseResultWriter.cs
@@ -1,8 +1,0 @@
-﻿using JTest.Core.Models;
-
-namespace JTest.Core.Output;
-
-public interface ITestCaseResultWriter
-{
-    void Write(TextWriter writer, JTestCaseResult testCaseResult, bool isDebug);
-}